refactor: 로그인 응답 반환 객체 리팩토링
